+++ src/idc/video/filter/__init__.py
@@ -1,3 +1,4 @@
 from ._drop_frames import DropFrames
 from ._filter_frames_by_label import FilterFramesByLabel
 from ._skip_similar_frames import SkipSimilarFrames
+from ._skip_similar_frames2 import SkipSimilarFrames2